# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tecnimotos API",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "EP-AUTH-01: Login con credenciales",
        "description": "Público — sin token.\nBloqueo 15 min tras 10 intentos fallidos/IP (07 §2.5).\nInMemory: rate limiting simplificado.",
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Login V1 Auth Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/mfa": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "EP-AUTH-02: Verificar MFA y obtener tokens",
        "description": "Semi-público — requiere mfa_session_token válido (TTL 5 min — 07 §2.4).\nSUPERADMIN/ADMINISTRADOR: código real enviado por correo (ADR-011).\nResto de roles: acepta cualquier código 6 dígitos (paso \"de forma\").\nRate limit: 20 req/min/IP vía RateLimiterMiddleware en /v1/auth/* (R31).\nCada intento se audita en mfa_intento (R29).",
        "operationId": "mfa_v1_auth_mfa_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Mfa V1 Auth Mfa Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "EP-AUTH-03: Rotar refresh token",
        "description": "Semi-público — requiere cookie refresh_token válido.\nReplay detection: reuso invalida familia completa (07 §2.3 Escenario 2).",
        "operationId": "refresh_v1_auth_refresh_post",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Refresh V1 Auth Refresh Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "EP-AUTH-04: Cerrar sesión",
        "description": "Todos autenticados — revoca la sesión y elimina cookie (07 §2.3 Escenario 1).",
        "operationId": "logout_v1_auth_logout_post",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Refresh Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Logout V1 Auth Logout Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/bootstrap-superadmin": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "EP-AUTH-06: Crear el primer SUPERADMIN",
        "description": "Público — sin token JWT.\nFunciona SOLO si no existe ningún SUPERADMIN en la base de datos.\nTras crear el primero, cualquier llamada posterior se rechaza permanentemente\n(la condición \"cero SUPERADMIN\" deja de cumplirse).\nLa clave de bootstrap proviene exclusivamente de SUPERADMIN_BOOTSTRAP_KEY (env).\nNunca se loguea la clave en texto plano.",
        "operationId": "bootstrap_superadmin_v1_auth_bootstrap_superadmin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapSuperadminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Bootstrap Superadmin V1 Auth Bootstrap Superadmin Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/registro": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "EP-AUTH-07: Autorregistro público con documentos",
        "description": "Público — sin token.\nCrea cuenta en PENDIENTE_DOCUMENTOS. No puede declarar ADMINISTRADOR/SUPERADMIN.\nMECANICO_MASTER/JUNIOR requieren certificado_tecnico adicional.\nPublica usuario.registro_pendiente para notificar al ADMINISTRADOR.",
        "operationId": "registro_v1_auth_registro_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_registro_v1_auth_registro_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Registro V1 Auth Registro Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_registro_v1_auth_registro_post": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 200,
            "minLength": 5,
            "title": "Email"
          },
          "nombre": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nombre"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "consentimiento_privacidad": {
            "type": "boolean",
            "title": "Consentimiento Privacidad"
          },
          "dni_frente": {
            "type": "string",
            "format": "binary",
            "title": "Dni Frente"
          },
          "dni_dorso": {
            "type": "string",
            "format": "binary",
            "title": "Dni Dorso"
          },
          "certificado_tecnico": {
            "anyOf": [
              {
                "type": "string",
                "format": "binary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certificado Tecnico"
          }
        },
        "type": "object",
        "required": [
          "email",
          "nombre",
          "password",
          "rol",
          "consentimiento_privacidad",
          "dni_frente",
          "dni_dorso"
        ],
        "title": "Body_registro_v1_auth_registro_post"
      },
...
      "BootstrapSuperadminRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 200,
            "minLength": 5,
            "title": "Email"
          },
          "nombre": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nombre"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password"
          },
          "bootstrap_key": {
            "type": "string",
            "minLength": 1,
            "title": "Bootstrap Key"
          }
        },
        "type": "object",
        "required": [
          "email",
          "nombre",
          "password",
          "bootstrap_key"
        ],
        "title": "BootstrapSuperadminRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "MfaRequest": {
        "properties": {
          "mfa_session_token": {
            "type": "string",
            "minLength": 1,
            "title": "Mfa Session Token"
          },
          "totp_code": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "title": "Totp Code"
          }
        },
        "type": "object",
        "required": [
          "mfa_session_token",
          "totp_code"
        ],
        "title": "MfaRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}